# [Xamarin.Android.Build.Tasks] <GetImportedLibraries/> perf improvements