@jonathanpeppers 2019-04-04
Build logs: [getimportedlibraries.zip](https://github.com/xamarin/xamarin-android/files/3045681/getimportedlibraries.zip)