@dellis1972 2019-04-08
3 of the errors look like those weird ssl issues. 

One is this

```
System.IO.FileNotFoundException: Could not find file "/Users/builder/jenkins/workspace/xamarin-android-pr-pipeline-release/xamarin-android/bin/TestRelease/TestResult-Mono.Android_TestsMultiDex.xml"
```

not sure it is related though,