@jonathanpeppers 2019-04-08
I think the macOS build is OK:

![image](https://user-images.githubusercontent.com/840039/55749974-e254f400-5a07-11e9-96a8-ab6af333379d.png)

They seem network-related, like the other PRs we have right now.